+++ app/src/main/java/com/futurice/freesound/arch/core/BaseApplication.java
@@ -14,29 +14,27 @@
  * limitations under the License.
  */
 
-package com.futurice.freesound.core;
+package com.futurice.freesound.arch.core;
 
 import com.futurice.freesound.inject.Injector;
 
-import android.os.Bundle;
+import android.app.Application;
 import android.support.annotation.CallSuper;
 import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 
 /**
- * A base Activity which provides a dependency injection mechanism.
+ * A base Application which provides a dependency injection mechanism.
  *
  * @param <T> The DI component class
  */
-public abstract class BaseActivity<T> extends AppCompatActivity implements Injector<T> {
+public abstract class BaseApplication<T> extends Application implements Injector<T> {
 
     private T component;
 
     @CallSuper
     @Override
-    protected void onCreate(@Nullable final Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
+    public void onCreate() {
+        super.onCreate();
         inject();
     }
 
@@ -51,4 +49,5 @@ public abstract class BaseActivity<T> extends AppCompatActivity implements Injec
 
     @NonNull
     protected abstract T createComponent();
+
 }
