+++ app/src/main/java/com/futurice/freesound/arch/core/BaseFragment.java
@@ -14,27 +14,29 @@
  * limitations under the License.
  */
 
-package com.futurice.freesound.core;
+package com.futurice.freesound.arch.core;
 
 import com.futurice.freesound.inject.Injector;
 
-import android.app.Application;
+import android.os.Bundle;
 import android.support.annotation.CallSuper;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
 
 /**
- * A base Application which provides a dependency injection mechanism.
+ * A base Fragment which provides a dependency injection mechanism.
  *
  * @param <T> The DI component class
  */
-public abstract class BaseApplication<T> extends Application implements Injector<T> {
+public abstract class BaseFragment<T> extends Fragment implements Injector<T> {
 
     private T component;
 
     @CallSuper
     @Override
-    public void onCreate() {
-        super.onCreate();
+    public void onActivityCreated(@Nullable final Bundle savedInstanceState) {
+        super.onActivityCreated(savedInstanceState);
         inject();
     }
