+++ app/src/main/java/com/futurice/freesound/arch/mvi/view/MviBaseFragment.java
@@ -18,7 +18,7 @@ package com.futurice.freesound.arch.mvi.view;
 
 import com.futurice.freesound.arch.mvi.Event;
 import com.futurice.freesound.arch.mvi.State;
-import com.futurice.freesound.core.BaseFragment;
+import com.futurice.freesound.arch.core.BaseFragment;
 
 import javax.inject.Inject;
